# contracts/openapi/api-gateway.yaml
openapi: 3.1.0
info:
  title: AI Code API Gateway
  version: 0.1.0
paths:
...
  /execution-plans/{id}/confirm:
    post:
      operationId: confirmExecutionPlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmExecutionPlanRequest"
      responses:
        "200":
          description: Execution result after confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTurnResponse"
  /execution-plans/{id}/reject:
    post:
      operationId: rejectExecutionPlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected execution plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlan"
  /execution-plans/{id}:
    get:
      operationId: getExecutionPlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Execution plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlan"
  /execution-ledger:
    get:
      operationId: getExecutionLedger
      parameters:
        - name: conversationId
          in: query
          required: false
          description: 仅返回指定会话产生的执行记录；不传时返回全局列表。
          schema:
            type: string
      responses:
        "200":
          description: Execution ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionLedgerItem"
...
  /expenses:
    get:
      operationId: getExpenses
      parameters:
        - name: conversationId
          in: query
          required: false
          description: 仅返回指定会话确认写入的费用记录；不传时返回全局列表。
          schema:
            type: string
      responses:
        "200":
          description: Expense records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseRecord"
  /expenses/{id}/submit:
    post:
      operationId: submitExpense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conversationId
          in: query
          required: true
          description: 直接 UI 提交的会话归属；后端会校验目标费用属于该会话并写入执行流水。
          schema:
            type: string
      responses:
        "200":
          description: Submitted expense draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conversationId
          in: query
          required: true
          description: 直接 UI 编辑的会话归属；后端会校验目标费用属于该会话并写入执行流水。
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseRecordUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
  /expenses/{id}/cancel:
    post:
      operationId: cancelExpense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conversationId
          in: query
          required: true
          description: 直接 UI 取消的会话归属；后端会校验目标费用属于该会话并写入执行流水。
          schema:
            type: string
      responses:
        "200":
          description: Canceled expense draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
  /reminders:
    get:
      operationId: getReminders
      parameters:
        - name: conversationId
          in: query
          required: false
          description: 仅返回指定会话确认写入的提醒；不传时返回全局列表。
          schema:
            type: string
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conversationId
          in: query
          required: true
          description: 直接 UI 编辑的会话归属；后端会校验目标提醒属于该会话并写入执行流水。
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}/complete:
    post:
      operationId: completeReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conversationId
          in: query
          required: true
          description: 直接 UI 完成的会话归属；后端会校验目标提醒属于该会话并写入执行流水。
          schema:
            type: string
      responses:
        "200":
          description: Completed reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}/cancel:
    post:
      operationId: cancelReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: conversationId
          in: query
          required: true
          description: 直接 UI 取消的会话归属；后端会校验目标提醒属于该会话并写入执行流水。
          schema:
            type: string
      responses:
        "200":
          description: Canceled reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
...
components:
  schemas:
...
    ConfirmExecutionPlanRequest:
      type: object
      required: [confirmToken]
      properties:
        confirmToken:
          type: string
        actionIds:
          type: array
          items:
            type: string
...
    DomainAction:
      type: object
      required: [id, planId, domain, actionType, status, riskLevel, summary, payload]
      properties:
        id:
          type: string
        planId:
          type: string
        domain:
          type: string
          enum: [calendar, expense, reminder]
        actionType:
          type: string
          enum: [calendar.cancel_event, calendar.create_event, calendar.query_events, calendar.update_event, expense.cancel_reimbursement, expense.create_reimbursement_draft, expense.submit_reimbursement, expense.update_reimbursement, reminder.cancel_reminder, reminder.complete_reminder, reminder.create_reminder, reminder.update_reminder]
        status:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        summary:
          type: string
        payload:
          type: object
        result:
          type: object
    ConfirmationCard:
      type: object
      required: [id, planId, status, requiredActionIds, title, description, confirmToken]
      properties:
        id:
          type: string
        planId:
          type: string
        status:
          type: string
        requiredActionIds:
          type: array
          items:
            type: string
        title:
          type: string
        description:
          type: string
        confirmToken:
          type: string
    ExecutionPlan:
      type: object
      required: [id, conversationId, status, riskLevel, summary, decisionTraceId, actions]
      properties:
        id:
          type: string
        conversationId:
          type: string
        status:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        summary:
          type: string
        decisionTraceId:
          type: string
        actions:
          type: array
          items:
            $ref: "#/components/schemas/DomainAction"
        confirmation:
          $ref: "#/components/schemas/ConfirmationCard"
    AgentTurnResponse:
      type: object
      required: [kind, conversationId]
      properties:
        kind:
          type: string
          enum: [assistant_message, clarification_request, confirmation_required, execution_result]
        conversationId:
          type: string
        message:
          type: string
        question:
          type: string
        clarificationId:
          type: string
        missingFields:
          type: array
          items:
            type: string
        quickReplies:
          type: array
          items:
            type: string
        quickReplyOptions:
          type: array
          items:
            $ref: "#/components/schemas/QuickReplyOption"
        plan:
          $ref: "#/components/schemas/ExecutionPlan"
        structuredElements:
          type: array
          items:
            type: object
    QuickReplyOption:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
...
    ExecutionLedgerItem:
      type: object
      required: [id, planId, eventType, status, message, createdAt]
      properties:
        id:
          type: string
        planId:
          type: string
        actionId:
          type: string
        eventType:
          type: string
        status:
          type: string
        message:
          type: string
        createdAt:
          type: string
...
    ExpenseRecord:
      type: object
      required: [id, title, currency, status, sourceActionId]
      properties:
        id:
          type: string
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        occurredOn:
          type: string
        status:
          type: string
          enum: [draft, submitted, canceled]
        sourceActionId:
          type: string
    ExpenseRecordUpdate:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        occurredOn:
          type: string
    Reminder:
      type: object
      required: [id, title, dueAt, status, sourceActionId]
      properties:
        id:
          type: string
        title:
          type: string
        dueAt:
          type: string
        status:
          type: string
          enum: [scheduled, done, canceled]
        sourceActionId:
          type: string
    ReminderUpdate:
      type: object
      properties:
        title:
          type: string
        dueAt:
          type: string